refactored search query validators into a separate file.
